+++ src/components/HeroSelect.js
@@ -30,35 +30,44 @@ class HeroSelect extends Component {
 
   render() {
     const { onToggle, theme } = this.props
+    const damageUnavailableClass = this.unavailableClass('damage')
+    const tankUnavailableClass = this.unavailableClass('tank')
+    const supportUnavailableClass = this.unavailableClass('support')
+    const damageAvailable = typeof damageUnavailableClass !== 'string'
+    const tankAvailable = typeof tankUnavailableClass !== 'string'
+    const supportAvailable = typeof supportUnavailableClass !== 'string'
 
     return (
       <div className="d-flex flex-wrap flex-justify-between">
         <div className="hero-box mb-3">
-          <h5 className={`h5 border-bottom pb-1 mb-2 ${this.unavailableClass('damage')}`}>
+          <h5 className={`h5 border-bottom pb-1 mb-2 ${damageUnavailableClass}`}>
             <RoleImage theme={theme} role="damage" size="12" className="d-inline-block mr-1" />
             Flankers
           </h5>
           <HeroCheckboxList
+            roleAvailable={damageAvailable}
             heroes={Hero.byType.Flanker}
             getUnavailableReason={this.getUnavailableReason}
             isChecked={this.isChecked}
             onToggle={onToggle}
           />
-          <h5 className={`h5 border-bottom pb-1 mt-4 mb-2 ${this.unavailableClass('tank')}`}>
+          <h5 className={`h5 border-bottom pb-1 mt-4 mb-2 ${tankUnavailableClass}`}>
             <RoleImage theme={theme} role="tank" size="12" className="d-inline-block mr-1" />
             Off-tanks
           </h5>
           <HeroCheckboxList
+            roleAvailable={tankAvailable}
             heroes={Hero.byType['Off-tank']}
             getUnavailableReason={this.getUnavailableReason}
             isChecked={this.isChecked}
             onToggle={onToggle}
           />
-          <h5 className={`h5 border-bottom pb-1 mt-4 mb-2 ${this.unavailableClass('support')}`}>
+          <h5 className={`h5 border-bottom pb-1 mt-4 mb-2 ${supportUnavailableClass}`}>
             <RoleImage theme={theme} role="support" size="12" className="d-inline-block mr-1" />
             Off-healers
           </h5>
           <HeroCheckboxList
+            roleAvailable={supportAvailable}
             heroes={Hero.byType['Off-healer']}
             getUnavailableReason={this.getUnavailableReason}
             isChecked={this.isChecked}
@@ -66,31 +75,34 @@ class HeroSelect extends Component {
           />
         </div>
         <div className="hero-box mb-3">
-          <h5 className={`h5 border-bottom pb-1 mb-2 ${this.unavailableClass('damage')}`}>
+          <h5 className={`h5 border-bottom pb-1 mb-2 ${damageUnavailableClass}`}>
             <RoleImage theme={theme} role="damage" size="12" className="d-inline-block mr-1" />
             Hitscan
           </h5>
           <HeroCheckboxList
+            roleAvailable={damageAvailable}
             heroes={Hero.byType.Hitscan}
             getUnavailableReason={this.getUnavailableReason}
             isChecked={this.isChecked}
             onToggle={onToggle}
           />
-          <h5 className={`h5 border-bottom pb-1 mt-4 mb-2 ${this.unavailableClass('tank')}`}>
+          <h5 className={`h5 border-bottom pb-1 mt-4 mb-2 ${tankUnavailableClass}`}>
             <RoleImage theme={theme} role="tank" size="12" className="d-inline-block mr-1" />
             Main tanks
           </h5>
           <HeroCheckboxList
+            roleAvailable={tankAvailable}
             heroes={Hero.byType['Main Tank']}
             getUnavailableReason={this.getUnavailableReason}
             isChecked={this.isChecked}
             onToggle={onToggle}
           />
-          <h5 className={`h5 border-bottom pb-1 mb-2 mt-4 ${this.unavailableClass('support')}`}>
+          <h5 className={`h5 border-bottom pb-1 mb-2 mt-4 ${supportUnavailableClass}`}>
             <RoleImage theme={theme} role="support" size="12" className="d-inline-block mr-1" />
             Main healers
           </h5>
           <HeroCheckboxList
+            roleAvailable={supportAvailable}
             heroes={Hero.byType['Main Healer']}
             getUnavailableReason={this.getUnavailableReason}
             isChecked={this.isChecked}
@@ -98,21 +110,23 @@ class HeroSelect extends Component {
           />
         </div>
         <div className="hero-box mb-3">
-          <h5 className={`h5 border-bottom pb-1 mb-2 ${this.unavailableClass('damage')}`}>
+          <h5 className={`h5 border-bottom pb-1 mb-2 ${damageUnavailableClass}`}>
             <RoleImage theme={theme} role="damage" size="12" className="d-inline-block mr-1" />
             DPS
           </h5>
           <HeroCheckboxList
+            roleAvailable={damageAvailable}
             heroes={Hero.byType.DPS}
             getUnavailableReason={this.getUnavailableReason}
             isChecked={this.isChecked}
             onToggle={onToggle}
           />
-          <h5 className={`h5 border-bottom pb-1 mt-4 mb-2 ${this.unavailableClass('damage')}`}>
+          <h5 className={`h5 border-bottom pb-1 mt-4 mb-2 ${damageUnavailableClass}`}>
             <RoleImage theme={theme} role="damage" size="12" className="d-inline-block mr-1" />
             Defense
           </h5>
           <HeroCheckboxList
+            roleAvailable={damageAvailable}
             heroes={Hero.byType.Defense}
             getUnavailableReason={this.getUnavailableReason}
             isChecked={this.isChecked}