+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed bug where role icons didn't change color when theme changed
 - Require role selection in match log form before you can set your new rank
+- Made it more clear which heroes are available based on selected role
 
 ## [Version 0.2.5](https://github.com/cheshire137/competiwatch-desktop/releases/tag/0.2.5)